% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filechoose.R
\name{getInfo}
\alias{getInfo}
\title{Get the information about the children's folder}
\usage{
getInfo(folderPath, dtoken)
}
\arguments{
\item{folderPath}{The path of the children's folder.}

\item{dtoken}{The Dropbox token generated by drop_auth (package rdrop2).}
}
\value{
returns a list of the size, mtime, atime, and ctime file.
}
\description{
This function works with fileGetterFile, fileGetterSave, fileGetterDir functions. 
When the last needs to get information about a children's folder they call this function, which look into
the the children's folder. Mtime of the children's folder is the minimum of all of its mtime files.
Atime of the children's folder is the minimum of all of its atime files. Ctime of the children's folder 
is the minimum of all of its ctime files. The size is a sum of its size files.
}
