% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filechoose.R, R/dirchoose.R, R/filesave.R
\name{shinyFilesDropBox-fileGetter}
\alias{shinyFilesDropBox-fileGetter}
\alias{fileGetterFile}
\alias{fileGetterDir}
\alias{fileGetterSave}
\title{Create a function that returns fileinfo according to the given restrictions}
\usage{
fileGetterFile(restrictions, filetypes, session, id, dtoken, roots = c(Home =
  ""))

fileGetterDir(restrictions, filetypes, session, id, dtoken, roots = c(Home =
  ""))

fileGetterSave(restrictions, filetypes, session, id, dtoken, roots = c(Home =
  ""))
}
\arguments{
\item{restrictions}{A vector of directories within the root that should be
filtered out of the results}

\item{filetypes}{A character vector of file extensions (without dot in front 
i.e. 'txt' not '.txt') to include in the output. Use the empty string to 
include files with no extension. If not set all file types will be included}

\item{dtoken}{The Dropbox token generated by drop_auth (package rdrop2).}

\item{roots}{A named vector of absolute filepaths or a function returning a 
named vector of absolute filepaths (the latter is useful if the volumes
should adapt to changes in the filesystem).}
}
\value{
A function taking a single path relative to the specified root, and
returns a list of files to be passed on to shiny
}
\description{
This functions returns a new function that can generate file information to 
be send to a shiny app based on a path relative to the given root. The 
function is secure in the sense that it prevents access to files outside of
the given root directory as well as to subdirectories matching the ones given
in restrictions. Furthermore can the output be filtered to only contain 
certain filetypes using the filter parameter.
}
