% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFilesDropBox-package.R
\docType{package}
\name{shinyFilesDropBox}
\alias{shinyFilesDropBox}
\alias{shinyFilesDropBox-package}
\title{shinyFilesDropBox package}
\description{
shinyFilesDropBox package allows you to navigate in your dropbox throught a dialog box. He is a sort of merging of the shinyFiles package and the rdrop2 package.
}
\author{
Armel Moth
}
